refactor: move protobuffers to code.domain
