+++ core/data/build.gradle.kts
@@ -12,9 +12,6 @@ plugins {
 //    alias(libs.plugins.google.services)
 
     alias(libs.plugins.travelinandroid.android.hilt)
-
-    // Proto DataStore
-    alias(libs.plugins.protobuf)
 }
 
 val localProperties = Properties().apply {
@@ -65,23 +62,4 @@ dependencies {
     // Proto DataStore
     implementation(libs.androidx.datastore)
     implementation(libs.protobuf.javalite)
-}
-
-protobuf {
-    protoc {
-        artifact = libs.protobuf.protoc.get().toString()
-    }
-
-    // Generates the java Protobuf-lite code for the Protobufs in this project. See
-    // https://github.com/google/protobuf-gradle-plugin#customizing-protobuf-compilation
-    // for more information.
-    generateProtoTasks {
-        all().configureEach {
-            builtins {
-                create("java") {
-                    option("lite")
-                }
-            }
-        }
-    }
 }
