+++ core/data/src/main/java/com/projectlab/core/data/di/NetworkModule.kt
@@ -7,11 +7,11 @@ import com.projectlab.core.data.config.AmadeusTokenSerializer
 import com.projectlab.core.data.network.AmadeusClientFactory
 import com.projectlab.core.data.network.AuthInterceptor
 import com.projectlab.core.data.network.HttpClientFactory
-import com.projectlab.core.data.proto.AmadeusToken
 import com.projectlab.core.data.remote.ActivitiesApiService
 import com.projectlab.core.data.remote.ActivityApiService
 import com.projectlab.core.data.remote.AmadeusApiService
 import com.projectlab.core.data.repository.AmadeusTokenProviderImpl
+import com.projectlab.core.domain.proto.AmadeusToken
 import com.projectlab.core.domain.repository.TokenProvider
 import dagger.Module
 import dagger.Provides
