+++ core/domain/src/main/proto/amadeus_token.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-option java_package = "com.projectlab.core.data.proto";
+option java_package = "com.projectlab.core.domain.proto";
 option java_multiple_files = true;
 
 message AmadeusToken {
